Get the set of available tasks

Retrieve a set of tasks that can be completed in Dr Frost Maths, choosing tasks to include and exclude.

Usage

dr_frost_tasks(include = NULL, exclude = NULL)

Arguments

include

A character vector of tasks to include. If NULL, all tasks will be included.

exclude

A character vector of tasks to exclude. If NULL, no tasks will be excluded.

x

A dr_frost_tasks object.

...

Not used.

Value

A data frame with additional class 'dr_frost_tasks', where each row describes a task. The table has 4 variables:

topic

The topic that the task falls under.

subtopic

The subtopic of the task.

task_name

The name of the task.

description

A brief description of the task.

Details

If the include argument matches a topic, subtopic or task name of any task, tasks will require a direct match with the string to be included. Otherwise, partial matching/pattern recognition will be used to identify matches.

The same applies to the exclude argument.

While this makes specifying tasks simple most of the time, it can sometimes lead to some unexpected cases where more complex searches result in less matches.

See also

Examples

dr_frost_tasks()
#> # 10 Dr Frost tasks
#> 
#> ── KS2 Number ──────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • addition_subtraction
#>   Add and subtract whole numbers.
#> • multiplication
#>   Multiply numbers up to 3 digit by 3 digit.
#> • pictoral_division
#>   Solve one step multiplication and division problems by using pictorial
#>   representations and arrays.
#> • division
#>   Divide numbers.
#> • number_facts
#>   Using number facts to solve connected calculations.
#> • missing_digits
#>   Solve multiplication, addition and subtraction problems involving missing
#>   digits.
#> • bidmas
#>   Understand the order in which operators in an expression are evaluated
#>   according to BIDMAS.
#> • estimate_calculations
#>   Estimate the result of a calculation by first rounding each number.
#> 
#> ── Times tables ────────────────────────────────────────────────────────────────
#> 
#> • fixed_time
#>   You have 30 seconds to answer as many questions as you can on all
#>   times/divide tables. Can you get on the leaderboard?
#> • individual_practice
#>   Practise each times table/divide table separately, with points for accuracy
#>   and speed.

dr_frost_tasks(include = "addition_subtraction")
#> # 1 Dr Frost task
#> 
#> ── KS2 Number ──────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • addition_subtraction
#>   Add and subtract whole numbers.

dr_frost_tasks(include = "Arithmetic Operations")
#> # 8 Dr Frost tasks
#> 
#> ── KS2 Number ──────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • addition_subtraction
#>   Add and subtract whole numbers.
#> • multiplication
#>   Multiply numbers up to 3 digit by 3 digit.
#> • pictoral_division
#>   Solve one step multiplication and division problems by using pictorial
#>   representations and arrays.
#> • division
#>   Divide numbers.
#> • number_facts
#>   Using number facts to solve connected calculations.
#> • missing_digits
#>   Solve multiplication, addition and subtraction problems involving missing
#>   digits.
#> • bidmas
#>   Understand the order in which operators in an expression are evaluated
#>   according to BIDMAS.
#> • estimate_calculations
#>   Estimate the result of a calculation by first rounding each number.

dr_frost_tasks(include = "KS3 Number")
#> # 0 Dr Frost tasks

dr_frost_tasks(include = "KS3", exclude = "Arithmetic")
#> # 0 Dr Frost tasks

dr_frost_tasks(exclude = "addition")
#> # 8 Dr Frost tasks
#> 
#> ── KS2 Number ──────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • multiplication
#>   Multiply numbers up to 3 digit by 3 digit.
#> • pictoral_division
#>   Solve one step multiplication and division problems by using pictorial
#>   representations and arrays.
#> • division
#>   Divide numbers.
#> • number_facts
#>   Using number facts to solve connected calculations.
#> • bidmas
#>   Understand the order in which operators in an expression are evaluated
#>   according to BIDMAS.
#> • estimate_calculations
#>   Estimate the result of a calculation by first rounding each number.
#> 
#> ── Times tables ────────────────────────────────────────────────────────────────
#> 
#> • fixed_time
#>   You have 30 seconds to answer as many questions as you can on all
#>   times/divide tables. Can you get on the leaderboard?
#> • individual_practice
#>   Practise each times table/divide table separately, with points for accuracy
#>   and speed.

tibble::as_tibble(dr_frost_tasks())
#> # A tibble: 10 × 4
#>    topic        subtopic                task_name             description       
#>    <chr>        <chr>                   <chr>                 <chr>             
#>  1 KS2 Number   "Arithmetic Operations" addition_subtraction  "Add and subtract…
#>  2 KS2 Number   "Arithmetic Operations" multiplication        "Multiply numbers…
#>  3 KS2 Number   "Arithmetic Operations" pictoral_division     "Solve one step m…
#>  4 KS2 Number   "Arithmetic Operations" division              "Divide numbers." 
#>  5 KS2 Number   "Arithmetic Operations" number_facts          " Using number fa…
#>  6 KS2 Number   "Arithmetic Operations" missing_digits        "Solve multiplica…
#>  7 KS2 Number   "Arithmetic Operations" bidmas                "Understand the o…
#>  8 KS2 Number   "Arithmetic Operations" estimate_calculations "Estimate the res…
#>  9 Times tables ""                      fixed_time            "You have 30 seco…
#> 10 Times tables ""                      individual_practice   "Practise each ti…