Get the set of available tasks

Retrieve a set of tasks that can be completed in Dr Frost Maths, choosing tasks to include and exclude.

Usage

dr_frost_tasks(include = NULL, exclude = NULL)

Arguments

include

A character vector of tasks to include. If NULL, all tasks will be included.

exclude

A character vector of tasks to exclude. If NULL, no tasks will be excluded.

Value

A data frame with additional class 'dr_frost_tasks', where each row describes a task. The table has 4 variables:

topic

The topic that the task falls under.

subtopic

The subtopic of the task.

task_name

The name of the task.

description

A brief description of the task.

Details

If the include argument matches a topic, subtopic or task name of any task, tasks will require a direct match with the string to be included. Otherwise, partial matching/pattern recognition will be used to identify matches.

The same applies to the exclude argument.

While this makes specifying tasks simple most of the time, it can sometimes lead to some unexpected cases where more complex searches result in less matches.

See also

Examples

dr_frost_tasks()
#> # 30 Dr Frost tasks
#> 
#> ── Algebra ─────────────────────────────────────────────────────────────────────
#> 
#> ── Sequences ──
#> • continue_sequence
#>   Continue a sequence.
#> • later_terms
#>   Find terms of a sequence given a term-to-term rule.
#> 
#> ── Solving Equations ──
#> • simple_substitution
#>   Substitute into simple expressions (limited to addition, subtraction,
#>   division, multiplication).
#> • solve_one_step
#>   Solve linear equations where the variable appears on one side of the equation
#>   only.
#> 
#> ── Data Handling & Probability ─────────────────────────────────────────────────
#> 
#> ── Averages and Range ──
#> • mean
#>   Calculate the mean as an average.
#> 
#> ── Data Representation ──
#> • pictograms
#>   Interpret a pictogram.
#> • bar_charts
#>   Bar Charts
#> • bank_statements
#>   Complete a bank statement.
#> • pie_charts
#>   Pie Charts
#> 
#> ── Number ──────────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • addition_subtraction
#>   Add and subtract whole numbers.
#> … with 20 more tasks
#> # ℹ Use `print(n = ...)` to see more tasks

dr_frost_tasks(include = "addition_subtraction")
#> # 1 Dr Frost task
#> 
#> ── Number ──────────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • addition_subtraction
#>   Add and subtract whole numbers.

dr_frost_tasks(include = "Arithmetic Operations")
#> # 8 Dr Frost tasks
#> 
#> ── Number ──────────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • addition_subtraction
#>   Add and subtract whole numbers.
#> • multiplication
#>   Multiply numbers up to 3 digit by 3 digit.
#> • pictoral_division
#>   Solve one step multiplication and division problems by using pictorial
#>   representations and arrays.
#> • division
#>   Divide numbers.
#> • number_facts
#>   Using number facts to solve connected calculations.
#> • missing_digits
#>   Solve multiplication, addition and subtraction problems involving missing
#>   digits.
#> • bidmas
#>   Understand the order in which operators in an expression are evaluated
#>   according to BIDMAS.
#> • estimate_calculations
#>   Estimate the result of a calculation by first rounding each number.

dr_frost_tasks(include = "Number")
#> # 19 Dr Frost tasks
#> 
#> ── Number ──────────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • addition_subtraction
#>   Add and subtract whole numbers.
#> • multiplication
#>   Multiply numbers up to 3 digit by 3 digit.
#> • pictoral_division
#>   Solve one step multiplication and division problems by using pictorial
#>   representations and arrays.
#> • division
#>   Divide numbers.
#> • number_facts
#>   Using number facts to solve connected calculations.
#> • missing_digits
#>   Solve multiplication, addition and subtraction problems involving missing
#>   digits.
#> • bidmas
#>   Understand the order in which operators in an expression are evaluated
#>   according to BIDMAS.
#> • estimate_calculations
#>   Estimate the result of a calculation by first rounding each number.
#> 
#> ── Decimals ──
#> • place_value
#>   Place value and ordering of decimals.
#> • decimal_addition_subtraction
#>   Add or subtract decimal numbers.
#> … with 9 more tasks
#> # ℹ Use `print(n = ...)` to see more tasks

dr_frost_tasks(include = "Number", exclude = "Arithmetic")
#> # 11 Dr Frost tasks
#> 
#> ── Number ──────────────────────────────────────────────────────────────────────
#> 
#> ── Decimals ──
#> • place_value
#>   Place value and ordering of decimals.
#> • decimal_addition_subtraction
#>   Add or subtract decimal numbers.
#> 
#> ── Fraction, Decimal & Percentage Correspondences ──
#> • conversion
#>   Convert between non-recurring decimals, fractions and percentages.
#> 
#> ── Fractions ──
#> • shape_fractions
#>   Find what fraction of a shape is shaded.
#> • equivalent_fractions
#>   Understand equivalent fractions.
#> • fraction_integer_division
#>   Divide fractions by integers.
#> • order_fractions
#>   Order fractions, possibly with decimals.
#> • mixed_addition_subtraction
#>   Add or subtract proper/improper fractions.
#> • fraction_amount
#>   Find a fraction of an amount.
#> • amount_before_fraction
#>   Find an original amount before a fraction of it was taken.
#> … with 1 more task
#> # ℹ Use `print(n = ...)` to see more tasks

dr_frost_tasks(exclude = "addition")
#> # 25 Dr Frost tasks
#> 
#> ── Algebra ─────────────────────────────────────────────────────────────────────
#> 
#> ── Sequences ──
#> • continue_sequence
#>   Continue a sequence.
#> • later_terms
#>   Find terms of a sequence given a term-to-term rule.
#> 
#> ── Solving Equations ──
#> • solve_one_step
#>   Solve linear equations where the variable appears on one side of the equation
#>   only.
#> 
#> ── Data Handling & Probability ─────────────────────────────────────────────────
#> 
#> ── Averages and Range ──
#> • mean
#>   Calculate the mean as an average.
#> 
#> ── Data Representation ──
#> • pictograms
#>   Interpret a pictogram.
#> • bar_charts
#>   Bar Charts
#> • bank_statements
#>   Complete a bank statement.
#> • pie_charts
#>   Pie Charts
#> 
#> ── Number ──────────────────────────────────────────────────────────────────────
#> 
#> ── Arithmetic Operations ──
#> • multiplication
#>   Multiply numbers up to 3 digit by 3 digit.
#> • pictoral_division
#>   Solve one step multiplication and division problems by using pictorial
#>   representations and arrays.
#> … with 15 more tasks
#> # ℹ Use `print(n = ...)` to see more tasks

tibble::as_tibble(dr_frost_tasks())
#> # A tibble: 30 × 4
#>    topic                       subtopic              task_name           descr…¹
#>    <chr>                       <chr>                 <chr>               <chr>  
#>  1 Algebra                     Sequences             continue_sequence   Contin…
#>  2 Algebra                     Sequences             later_terms         Find t…
#>  3 Algebra                     Solving Equations     simple_substitution Substi…
#>  4 Algebra                     Solving Equations     solve_one_step      Solve …
#>  5 Data Handling & Probability Averages and Range    mean                Calcul…
#>  6 Data Handling & Probability Data Representation   pictograms          Interp…
#>  7 Data Handling & Probability Data Representation   bar_charts          Bar Ch…
#>  8 Data Handling & Probability Data Representation   bank_statements     Comple…
#>  9 Data Handling & Probability Data Representation   pie_charts          Pie Ch…
#> 10 Number                      Arithmetic Operations addition_subtracti… Add an…
#> # … with 20 more rows, and abbreviated variable name ¹​description