Complete Dr Frost Maths tasks

Complete a set of Dr Frost Maths tasks. Each task rewards the user with a certain number of points, which can vary.

Usage

perform_tasks(
  tasks = dr_frost_tasks(),
  email = NULL,
  password = NULL,
  keyring = NULL,
  end = TRUE
)

Arguments

tasks

The tasks to complete. Either a dr_frost_tasks() object, or a character vector. If a character vector is supplied, it will be passed as the include argument into the dr_frost_tasks() function. By default, all available tasks are completed.

email

Used to login to the website. If NULL, the default email is used.

password

Used to login to the website. If NULL, the function will try to find the password corresponding to the email that has been set using dr_frost_account().

keyring

A string, passed into keyring::key_get(). Use this if your Dr Frost credentials are stored on a keyring that is not the default.

end

Whether the browser session should be closed when the tasks are completed.

Value

A data frame of information about each task, with 4 variables:

task

The task that was attempted.

completed

Whether the task was successfully completed.

points

The number of points that were earned by completing the task.

error

If the task failed, the error object that was returned.

Details

Completing Dr Frost Tasks requires a continuous (although not necessarily fast) internet connection, and tasks will fail if no connection is found for an extended period of time.

Completing each task calls a corresponding function in Python.

See also

Examples

if (FALSE) {
  perform_tasks("addition")
  perform_tasks()
}